% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_label}
\alias{add_label}
\title{Add a label to a vector}
\usage{
add_label(x, label)
}
\arguments{
\item{x}{Vector to add label to}

\item{label}{Label to add to vector}
}
\value{
Labeled vector
}
\description{
Add a label to a vector
}
