% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_labels}
\alias{add_labels}
\title{Add Labels to Dataframe}
\usage{
add_labels(.data, ...)
}
\arguments{
\item{.data}{Dataframe that you want to add labels to}

\item{...}{Labeled vector of labels, with the name of the element equal to
the name of the column and the element equaling the label, or comma-seperated name-value pair}
}
\value{
Dataframe with labels
}
\description{
Add Labels to Dataframe
}
