% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{check_columns}
\alias{check_columns}
\title{Check all data frames include the correct types of columns}
\usage{
check_columns(
  ds_spec = NULL,
  ds_vars = NULL,
  var_spec = NULL,
  value_spec = NULL,
  derivations = NULL,
  codelist = NULL,
  supp = NULL
)
}
\arguments{
\item{ds_spec}{dataset specification}

\item{ds_vars}{dataset variables}

\item{var_spec}{variable specification}

\item{value_spec}{value specification}

\item{derivations}{derivation information}

\item{codelist}{codelist information}

\item{supp}{supp information}
}
\description{
This function checks for vector types and accepted words
}
