% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{check_DatasetMeta}
\alias{check_DatasetMeta}
\title{Check that the Class Type of an object is DatasetMeta.}
\usage{
check_DatasetMeta(x)
}
\arguments{
\item{x}{An object whose class type needs to be checked.}
}
\value{
Logical: TRUE if the class type of \code{metacore} is \code{DatasetMeta},
otherwise abort with errors.
}
\description{
This is an internal function that is a wrapper to the exported functions
\code{is_metacore} and \code{is_DatasetMeta}.

This function is used as a guard clause in many features of the {metatools}
package that are intended only to be used with the subsetted Metacore object
of class type DatasetMeta. If either of the wrapped functions return \code{FALSE }
then an appropriate error message is displayed.
}
\examples{
load(metacore_example("pilot_ADaM.rda"))
dm <- select_dataset(metacore, "DM", quiet = TRUE)
check_DatasetMeta("DUMMY")   # Expect error
check_DatasetMeta(metacore)  # Expect error
check_DatasetMeta(dm)        # Expect valid, i.e., return TRUE
}
