% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_words}
\alias{check_words}
\title{Check Words in Column}
\usage{
check_words(..., col)
}
\arguments{
\item{...}{permissible words in the column}

\item{col}{the column to check for specific words}
}
\description{
Check Words in Column
}
