% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_builder.R
\name{create_supp_table}
\alias{create_supp_table}
\title{Create supp table}
\usage{
create_supp_table(
  doc,
  cols = c(dataset = "[D|d]ataset|[D|d]omain", where = "[W|w]here [C|clause]", type =
    "[T|t]ype", length = "[L|l]ength", origin = "[O|o]rigin"),
  sheet = NULL,
  where_sep_sheet = TRUE,
  where_cols = c(id = "ID", variable = "[V|v]ariable", comparator = "[C|c]omparator",
    value = "[V|v]alue"),
  var_spec = NULL,
  value_spec = NULL,
  codelist = NULL,
  comments = NULL
)
}
\arguments{
\item{value_spec}{value_spec table from metacore object}

\item{codelist}{codelist table from metacore object}

\item{comments}{comments table from metacore object (optional)}
}
\value{
a dataset formatted for the metacore object (supp table)
}
\description{
Creates the supp table from value_spec, codelist, and comments by identifying
supplemental datasets (SUPP*) and extracting their metadata. For each SUPP domain,
extracts variable names from the QNAM codelist, identifying variables from the
IDVAR comment, and evaluator from the QEVAL codelist.
}
\details{
Note for future: length is populated only when VLM exists
}
\seealso{
Other spec builders: 
\code{\link{spec_type_to_codelist}()},
\code{\link{spec_type_to_comments}()},
\code{\link{spec_type_to_derivations}()},
\code{\link{spec_type_to_ds_spec}()},
\code{\link{spec_type_to_ds_vars}()},
\code{\link{spec_type_to_value_spec}()},
\code{\link{spec_type_to_var_spec}()}
}
\concept{spec builders}
