% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDef.R
\name{DataDef_print}
\alias{DataDef_print}
\title{DataDef class print function}
\usage{
DataDef_print(...)
}
\arguments{
\item{...}{pass in the dataframes to be validated}
}
\description{
DataDef class print function
}
\seealso{
Other DataDef: 
\code{\link{DataDef_initialize}()},
\code{\link{DataDef_validate}()},
\code{\link{DataDef}},
\code{\link{datadef}()},
\code{\link{readonly}()}
}
\concept{DataDef}
