% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDef.R
\name{DataDef_validate}
\alias{DataDef_validate}
\title{DataDef R6 object validation function}
\usage{
DataDef_validate()
}
\description{
This checks that the labels and lengths of ds_vars match var_spec
}
\seealso{
Other DataDef: 
\code{\link{DataDef_initialize}()},
\code{\link{DataDef_print}()},
\code{\link{DataDef}},
\code{\link{datadef}()},
\code{\link{readonly}()}
}
\concept{DataDef}
