% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_builders.R
\name{define_to_DataDef}
\alias{define_to_DataDef}
\title{Define XML to DataDef Object}
\usage{
define_to_DataDef(path)
}
\arguments{
\item{path}{loaction of the define xml as a string}
}
\value{
DataDef Object
}
\description{
Given a path, this function converts the define xml to a DataDef Object
}
