% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacore.R
\name{filter_dataset}
\alias{filter_dataset}
\title{Filter metacore object to single dataset}
\usage{
filter_dataset(.data, dataset)
}
\arguments{
\item{.data}{the metacore object of dataframes}

\item{dataset}{the specific dataset to subset by}
}
\value{

}
\description{
Filter metacore object to single dataset
}
