% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metacore.R
\name{get_keys}
\alias{get_keys}
\title{Get Dataset Keys}
\usage{
get_keys(metacode, dataset)
}
\arguments{
\item{metacode}{metacore object}

\item{dataset}{A dataset name}
}
\value{
a 2-column tibble with dataset key variables and key sequence
}
\description{
Returns the dataset keys for a given dataset
}
\examples{
\dontrun{
meta_ex <- spec_to_metacore(metacore_example("p21_mock.xlsx"))
get_keys(meta_ex, "AE")
get_keys(meta_ex, AE)
}
}
