% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{is_DatasetMeta}
\alias{is_DatasetMeta}
\title{Is DatasetMeta object}
\usage{
is_DatasetMeta(x)
}
\arguments{
\item{x}{object to check}
}
\value{
\code{TRUE} if DatasetMeta, \code{FALSE} if not
}
\description{
Is DatasetMeta object
}
\examples{
load(metacore_example("pilot_ADaM.rda"))
dm <- select_dataset(metacore, "DM", quiet = TRUE)
if_DatasetMeta("DUMMY")   # Expect error
is_DatasetMeta(metacore)  # Expect error
is_DatasetMeta(dm)        # Expect valid, i.e., return TRUE
}
