% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{is_metacore}
\alias{is_metacore}
\title{Is metacore object}
\usage{
is_metacore(x)
}
\arguments{
\item{x}{object to check}
}
\value{
\code{TRUE} if metacore, \code{FALSE} if not
}
\description{
Is metacore object
}
\examples{
# Loads in a metacore obj called metacore
load(metacore_example("pilot_ADaM.rda"))
is_metacore(metacore)

}
