% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{metacore_example}
\alias{metacore_example}
\title{Get path to metacore example}
\usage{
metacore_example(file = NULL)
}
\arguments{
\item{file}{Name of file. If \code{NULL}, the example files will be listed.}
}
\description{
metacore comes bundled with a number of sample files in its \code{inst/extdata}
directory. This function make them easy to access. When testing or writing
examples in other packages, it is best to use the 'pilot_ADaM.rda' example as
it loads fastest.
}
\examples{
metacore_example()
metacore_example("mock_spec.xlsx")
}
