% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quiet_if_true}
\alias{quiet_if_true}
\title{Conditionally suppress messages and warnings}
\usage{
quiet_if_true(expr, quiet = FALSE)
}
\description{
Conditionally suppress messages and warnings
}
\keyword{internal}
