% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_builder.R
\name{read_all_sheets}
\alias{read_all_sheets}
\title{Read in all Sheets}
\usage{
read_all_sheets(path)
}
\arguments{
\item{path}{string of the file path}
}
\value{
a list of datasets
}
\description{
Given a path to a file, this function reads in all sheets of an excel file
}
