% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_builder.R
\name{spec_type}
\alias{spec_type}
\title{Check the type of spec document}
\usage{
spec_type(path)
}
\arguments{
\item{path}{file location as a string}
}
\value{
returns string indicating the type of spec document
}
\description{
Check the type of spec document
}
