% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_builder.R
\name{spec_type_to_supp}
\alias{spec_type_to_supp}
\title{Spec to supp}
\usage{
spec_type_to_supp(
  doc,
  cols = c(dataset = "[D|d]ataset|[D|d]omain", variable = "[V|v]ariable", where =
    "[W|w]here [C|clause]", type = "[T|t]ype", length = "[L|l]ength", origin =
    "[O|o]rigin"),
  sheet = NULL,
  where_sep_sheet = TRUE,
  where_cols = c(id = "ID", variable = "[V|v]ariable", comparator = "[C|c]omparator",
    value = "[V|v]alue"),
  var_spec = NULL,
  value_spec = NULL,
  codelist = NULL,
  comments = NULL
)
}
\arguments{
\item{doc}{Named list of datasets @seealso \code{\link[=read_all_sheets]{read_all_sheets()}} for exact format}

\item{cols}{Named vector of column names. The column names can be regular
expressions for more flexibility. But, the names must follow the given pattern.
For modern Pinnacle 21 specification spec types, the mapping will typically be
the same as the mapping for the \code{value_spec} table.}

\item{sheet}{Regular expression for the sheet name}

\item{where_sep_sheet}{Boolean value to control if the where information in a
separate dataset. If the where information is on a separate sheet, set to
true and provide the column information with the \code{where_cols} inputs.}

\item{where_cols}{Named list with an id and where field. All columns in the
where field will be collapsed together}

\item{var_spec}{var_spec table from the metacore object}

\item{value_spec}{value_spec table from the metacore object}

\item{codelist}{codelist table from metacore object}

\item{comments}{comments table from metacore object (optional). Only required when
generating a define.xml enabled metacore object.}
}
\value{
a dataset formatted for the metacore object (supp table)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Creates the supp table from value_spec, codelist, and comments by identifying
supplemental datasets (SUPP*) and extracting their metadata. For each SUPP domain,
extracts variable names from the QNAM codelist, identifying variables from the
IDVAR comment, and evaluator from the QEVAL codelist.
}
\details{
Note for future: length is populated only when VLM exists
}
\seealso{
Other spec builders:
\code{\link[=spec_type_to_codelist]{spec_type_to_codelist()}},
\code{\link[=spec_type_to_comments]{spec_type_to_comments()}},
\code{\link[=spec_type_to_derivations]{spec_type_to_derivations()}},
\code{\link[=spec_type_to_documents]{spec_type_to_documents()}},
\code{\link[=spec_type_to_ds_spec]{spec_type_to_ds_spec()}},
\code{\link[=spec_type_to_ds_vars]{spec_type_to_ds_vars()}},
\code{\link[=spec_type_to_value_spec]{spec_type_to_value_spec()}},
\code{\link[=spec_type_to_var_spec]{spec_type_to_var_spec()}}
}
\concept{spec builders}
