% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_builders.R
\name{xml_to_comments}
\alias{xml_to_comments}
\title{XML to comments table}
\usage{
xml_to_comments(doc)
}
\arguments{
\item{doc}{xml document}
}
\value{
dataframe with comment_id's and comment text
}
\description{
Reads def:CommentDef elements from an XML document and builds a comments
table. Comments are linked to variables via value_spec.comment_id.
}
\seealso{
Other xml builder: 
\code{\link{xml_to_codelist}()},
\code{\link{xml_to_derivations}()},
\code{\link{xml_to_ds_spec}()},
\code{\link{xml_to_ds_vars}()},
\code{\link{xml_to_value_spec}()},
\code{\link{xml_to_var_spec}()}
}
\concept{xml builder}
