% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_builders.R
\name{xml_to_ds_vars}
\alias{xml_to_ds_vars}
\title{XML to Data Set Var table}
\usage{
xml_to_ds_vars(doc)
}
\arguments{
\item{doc}{xml document}
}
\value{
data frame with the dataset and variables
}
\description{
Creates the ds_vars table, which acts as a key between the datasets and the var spec
}
\seealso{
Other xml builder:
\code{\link[=xml_to_codelist]{xml_to_codelist()}},
\code{\link[=xml_to_derivations]{xml_to_derivations()}},
\code{\link[=xml_to_ds_spec]{xml_to_ds_spec()}},
\code{\link[=xml_to_value_spec]{xml_to_value_spec()}},
\code{\link[=xml_to_var_spec]{xml_to_var_spec()}}
}
\concept{xml builder}
