% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{check_columns}
\alias{check_columns}
\title{Check all data frames include the correct types of columns}
\usage{
check_columns(ds_spec, ds_vars, var_spec, value_spec, derivations, code_list)
}
\description{
This function checks for vector types and accepted words
}
