% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_builder.R
\name{create_tbl}
\alias{create_tbl}
\title{Create table}
\usage{
create_tbl(doc, cols, optional = NULL, context = NULL)
}
\arguments{
\item{doc}{list of sheets from a excel doc}

\item{cols}{vector of regex to get a datasets base on which columns it has.
If the vector is named it will also rename the columns}

\item{context}{Provides the calling context for better error messaging to the user}
}
\value{
dataset (or list of datasets if not specific enough)
}
\description{
This function creates a table from excel sheets. This is mainly used
internally for building spec readers, but is exported so others who need to
build spec readers can use it.
}
