% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_builder.R
\name{create_tbl}
\alias{create_tbl}
\title{Select sheet}
\usage{
create_tbl(doc, cols)
}
\arguments{
\item{doc}{list of sheets from a excel dos}

\item{cols}{vector of regex to get a datasets base on which columns it has.
If the vector is named it will also rename the columns}
}
\value{
dataset (or list of datasets if not specific enough)
}
\description{
Helper function to select a sheet when building a spec reader. Creates a tbl based on
a provided excel document and specified column names.
}
