% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDef.R
\name{DataDef}
\alias{DataDef}
\title{The DataDef R6 Class}
\arguments{
\item{ds_spec}{things and stuff}

\item{ds_vars}{things and stuff}

\item{var_spec}{things and stuff}

\item{value_spec}{things and stuff}

\item{derivations}{things and stuff}

\item{code_list}{things and stuff}

\item{...}{things and stuff}
}
\description{
This uses the initialize, print, and validate functions above to create a single object
The user can query
}
\section{Fields}{

\describe{
\item{\code{ds_spec}}{First or full name of the person.}

\item{\code{ds_vars}}{First or full name of the person.}

\item{\code{var_spec}}{First or full name of the person.}

\item{\code{value_spec}}{First or full name of the person.}

\item{\code{derivations}}{First or full name of the person.}

\item{\code{codelist}}{First or full name of the person.}

\item{\code{changelog}}{First or full name of the person.}
}}

