% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDef_helpers.R
\name{DataDef_initialize}
\alias{DataDef_initialize}
\title{DataDef class initilization function}
\usage{
DataDef_initialize(
  ds_spec,
  ds_vars,
  var_spec,
  value_spec,
  derivations,
  code_list
)
}
\arguments{
\item{ds_spec}{contians each dataset in the study, with the labels for each}

\item{ds_vars}{information on what variables are in each dataset + plus dataset specific variable information}

\item{var_spec}{variable information that is shared across all datasets}

\item{value_spec}{parameter specific information, as data is long the specs for wbc might be difference the hgb}

\item{derivations}{contains derivation, it allows for different variables to have the same derivation}

\item{code_list}{contains the code/decode information}
}
\description{
This function assigns the datasets to readonly object
and runs validation on those objects
}
