% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDef.R
\name{DataDef_initialize}
\alias{DataDef_initialize}
\title{This file includes the internal functions needed to create the readonly DataDef R6 object}
\usage{
DataDef_initialize(
  ds_spec,
  ds_vars,
  var_spec,
  value_spec,
  derivations,
  code_list
)
}
\description{
This file includes the internal functions needed to create the readonly DataDef R6 object
}
