% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_builders.R
\name{define_to_metacore}
\alias{define_to_metacore}
\title{Define XML to DataDef Object}
\usage{
define_to_metacore(path, quiet = deprecated(), verbose = "message")
}
\arguments{
\item{path}{location of the define xml as a string}

\item{quiet}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}} Option to quietly load in, this
will suppress warnings, but not errors. Expects either \code{TRUE} or \code{FALSE}.
Default behaviour is \code{FALSE}. As of v0.3.0 this argument is deprecated in favour
of \code{verbose}.}

\item{verbose}{A character string specifying the desired verbosity level.
Must be one of:
\describe{
\item{"message"}{ (default) Messages and warnings are handled normally.}
\item{"warn"}{Messages are suppressed, but warnings are allowed.}
\item{"collapse"}{Warnings are collapsed into a single message indicating the
number of suppressed warnings.}
\item{"silent"}{Both messages and warnings are suppressed.}
}}
}
\value{
Metacore/DataDef object
}
\description{
Given a path, this function converts the define xml to a DataDef/Metacore object.
}
