% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_builders.R
\name{define_to_metacore}
\alias{define_to_metacore}
\title{Define XML to metacore Object}
\usage{
define_to_metacore(path)
}
\arguments{
\item{path}{loaction of the define xml as a string}
}
\value{
metacore Object
}
\description{
Given a path, this function converts the define xml to a metacore Object
}
