% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_builders.R
\name{define_to_metacore}
\alias{define_to_metacore}
\title{Define XML to DataDef Object}
\usage{
define_to_metacore(path, quiet = FALSE)
}
\arguments{
\item{path}{location of the define xml as a string}

\item{quiet}{Option to quietly load in; when \code{TRUE}, messages and warnings
are suppressed, but errors are still raised.}
}
\value{
Metacore/DataDef object
}
\description{
Given a path, this function converts the define xml to a DataDef/Metacore object.
}
