% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{is_DatasetMeta}
\alias{is_DatasetMeta}
\title{Is DatasetMeta object}
\usage{
is_DatasetMeta(x)
}
\arguments{
\item{x}{object to check}
}
\value{
\code{TRUE} if DatasetMeta, \code{FALSE} if not
}
\description{
Is DatasetMeta object
}
\examples{
load(metacore_example("pilot_ADaM.rda"))
adsl <- select_dataset(metacore, "ADSL", quiet = TRUE)
is_DatasetMeta("DUMMY")   # Expect FALSE
is_DatasetMeta(metacore)  # Expect FALSE
is_DatasetMeta(adsl)      # Expect TRUE
}
