% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacore.R
\name{load_metacore}
\alias{load_metacore}
\title{load metacore object}
\usage{
load_metacore(path = NULL)
}
\arguments{
\item{path}{location of the metacore object to load into memory}
}
\value{
metacore object in memory
}
\description{
load metacore object
}
