% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacore-package.R
\docType{package}
\name{metacore-package}
\alias{metacore-package}
\title{metacore: A Centralized Metadata Object Focus on Clinical Trial Data Programming Workflows}
\description{
Create an immutable container holding metadata for the purpose of better enabling programming activities and functionality of other packages within the clinical programming workflow.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://atorus-research.github.io/metacore/}
  \item \url{https://github.com/atorus-research/metacore}
  \item Report bugs at \url{https://github.com/atorus-research/metacore/issues}
}

}
\author{
\strong{Maintainer}: Liam Hobby \email{liam.f.hobby@gsk.com}

Authors:
\itemize{
  \item Christina Fillmore \email{christina.e.fillmore@gsk.com} (\href{https://orcid.org/0000-0003-0595-2302}{ORCID})
  \item Bill Denney
  \item Maya Gans \email{maya.gans@atorusresearch.com} (\href{https://orcid.org/0000-0002-5452-6089}{ORCID})
  \item Ashley Tarasiewicz \email{ashley.tarasiewicz@atorusresearch.com}
  \item Mike Stackhouse \email{mike.stackhouse@atorusresearch.com} (\href{https://orcid.org/0000-0001-6030-723X}{ORCID})
  \item Tamara Senior \email{tamara.senior@roche.com}
}

Other contributors:
\itemize{
  \item GSK/Atorus JPT [copyright holder, funder]
}

}
\concept{internal}
\keyword{internal}
