% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacore.R
\name{MetaCore_filter}
\alias{MetaCore_filter}
\title{Select method to subset by a single dataframe}
\usage{
MetaCore_filter(value)
}
\arguments{
\item{value}{the dataframe to subset by}
}
\description{
Select method to subset by a single dataframe
}
