% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDef_helpers.R
\name{readonly}
\alias{readonly}
\title{readonly function factory}
\usage{
readonly(name)
}
\arguments{
\item{name}{the name of the readonly object}

\item{value}{any attempt at assignment to the readonly object}
}
\description{
This function is used inside the R6 active method and allows us
to read the selected dataframe and prevents overwriting
}
