% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacore.R
\name{save_metacore}
\alias{save_metacore}
\title{save metacore object}
\usage{
save_metacore(metacore_object)
}
\arguments{
\item{metacore_object}{the metacore object in memory to save to disc}
}
\value{
an .rda file with the same name as your object in memory saved to your working directory
}
\description{
save metacore object
}
