% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacore.R
\name{save_metacore}
\alias{save_metacore}
\title{save metacore object}
\usage{
save_metacore(metacore_object, path = NULL)
}
\arguments{
\item{metacore_object}{the metacore object in memory to save to disc}

\item{path}{file path and file name to save metacore object}
}
\value{
an .rda file
}
\description{
save metacore object
}
