% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacore.R
\name{select_dataset}
\alias{select_dataset}
\title{Select metacore object to single dataset}
\usage{
select_dataset(
  .data,
  dataset,
  simplify = FALSE,
  quiet = deprecated(),
  verbose = "message"
)
}
\arguments{
\item{.data}{the metacore object of dataframes}

\item{dataset}{the specific dataset to subset by}

\item{simplify}{return a single dataframe}

\item{quiet}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}} Option to quietly load in, this
will suppress warnings, but not errors. Expects either \code{TRUE} or \code{FALSE}.
Default behaviour is \code{FALSE}. As of v0.3.0 this argument is deprecated in favour
of \code{verbose}.}

\item{verbose}{A character string specifying the desired verbosity level.
Must be one of:
\describe{
\item{"message"}{(default) Messages and warnings are handled normally.}
\item{"warn"}{Messages are suppressed, but warnings are allowed.}
\item{"collapse"}{Warnings are collapsed into a single message indicating the
number of suppressed warnings.}
\item{"silent"}{Both messages and warnings are suppressed.}
}}
}
\value{
a filtered subset of the metacore object
}
\description{
Select metacore object to single dataset
}
