% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_builder.R
\name{spec_to_metacore}
\alias{spec_to_metacore}
\title{Specification document to metacore object}
\usage{
spec_to_metacore(
  path,
  quiet = deprecated(),
  where_sep_sheet = TRUE,
  verbose = "message"
)
}
\arguments{
\item{path}{string of file location}

\item{quiet}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}} Option to quietly load in, this
will suppress warnings, but not errors. Expects either \code{TRUE} or \code{FALSE}.
Default behaviour is \code{FALSE}.}

\item{where_sep_sheet}{Option to tell if the where is in a separate sheet,
like in older p21 specs or in a single sheet like newer p21 specs.}

\item{verbose}{A character string specifying the desired verbosity level.
Must be one of:
\describe{
\item{"message"}{ (default) Messages and warnings are handled normally.}
\item{"warn"}{Messages are suppressed, but warnings are allowed.}
\item{"silent"}{Both messages and warnings are suppressed.}
}}
}
\value{
given a spec document it returns a metacore object
}
\description{
This function takes the location of an excel specification document and reads
it in as a meta core object. At the moment it only supports specification in
the format of pinnacle 21 specifications. But, the section level spec builder can
be used as building blocks for bespoke specification documents.
}
