% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_builder.R
\name{spec_to_metacore}
\alias{spec_to_metacore}
\title{Specification document to metacore object}
\usage{
spec_to_metacore(path, quiet = FALSE, where_sep_sheet = TRUE)
}
\arguments{
\item{path}{string of file location}

\item{quiet}{Option to quietly load in, this will suppress warnings, but not
errors}

\item{where_sep_sheet}{Option to tell if the where is in a separate sheet,
like in older p21 specs or in a single sheet like newer p21 specs}
}
\value{
given a spec document it returns a metacore object
}
\description{
This function takes the location of an excel specification document and reads
it in as a meta core object. At the moment it only supports specification in
the format of pinnacle 21 specifications. But, the @family spec builder can
be used as building blocks for bespoke specification documents
}
