% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_builder.R
\name{spec_type_to_derivations}
\alias{spec_type_to_derivations}
\title{Spec to derivation}
\usage{
spec_type_to_derivations(
  doc,
  cols = c(derivation_id = "ID", derivation = "[D|d]efinition|[D|d]escription",
    method_name = "[N|n]ame", method_type = "[T|t]ype", document_id = "[D|d]ocument",
    pages = "[P|p]ages"),
  sheet = "Method|Derivations?",
  var_cols = c(dataset = "[D|d]ataset|[D|d]omain", variable = "[N|n]ame|[V|v]ariables?",
    origin = "[O|o]rigin", predecessor = "[P|p]redecessor", comment = "[C|c]omment")
)
}
\arguments{
\item{doc}{Named list of datasets @seealso \code{\link[=read_all_sheets]{read_all_sheets()}} for exact
format}

\item{cols}{Named vector of column names. The column names can be regular
expressions for more flexibility. But, the names must follow the given
pattern}

\item{sheet}{Regular expression for the sheet name}

\item{var_cols}{Named vector of the name(s) of the origin, predecessor and
comment columns. These do not have to be on the specified sheet.}
}
\value{
a dataset formatted for the metacore object
}
\description{
Creates the derivation table from a list of datasets (optionally filtered by
the sheet input). The named vector \code{cols} is used to determine which is the
correct sheet and renames the columns. The derivation will be used for
"derived" origins, the comments for "assigned" origins, and predecessor for
"predecessor" origins.
}
\seealso{
Other spec builders: 
\code{\link{create_supp_table}()},
\code{\link{spec_type_to_codelist}()},
\code{\link{spec_type_to_comments}()},
\code{\link{spec_type_to_ds_spec}()},
\code{\link{spec_type_to_ds_vars}()},
\code{\link{spec_type_to_value_spec}()},
\code{\link{spec_type_to_var_spec}()}
}
\concept{spec builders}
