% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_builder.R
\name{spec_type_to_ds_spec}
\alias{spec_type_to_ds_spec}
\title{Spec to ds_spec}
\usage{
spec_type_to_ds_spec(
  doc,
  cols = c(dataset = "[N|n]ame|[D|d]ataset|[D|d]omain", structure = "[S|s]tructure",
    label = "[L|l]abel|[D|d]escription", class = "[C|c]lass", repeating =
    "[R|r]epeating", reference = "[R|r]eference [D|d]ata", purpose = "[P|p]urpose"),
  sheet = NULL
)
}
\arguments{
\item{doc}{Named list of datasets @seealso \code{\link[=read_all_sheets]{read_all_sheets()}} for exact
format}

\item{cols}{Named vector of column names. The column names can be regular
expressions for more flexibility. But, the names must follow the given pattern}

\item{sheet}{Regular expression for the sheet name}
}
\value{
a dataset formatted for the metacore object
}
\description{
Creates the ds_spec from a list of datasets (optionally filtered by the sheet
input). The named vector \code{cols} is used to determine which is the correct
sheet and renames the columns
}
\seealso{
Other spec builders: 
\code{\link{create_supp_table}()},
\code{\link{spec_type_to_codelist}()},
\code{\link{spec_type_to_comments}()},
\code{\link{spec_type_to_derivations}()},
\code{\link{spec_type_to_ds_vars}()},
\code{\link{spec_type_to_value_spec}()},
\code{\link{spec_type_to_var_spec}()}
}
\concept{spec builders}
