% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_builder.R
\name{spec_type_to_ds_vars}
\alias{spec_type_to_ds_vars}
\title{Spec to ds_vars}
\usage{
spec_type_to_ds_vars(
  doc,
  cols = c(dataset = "[D|d]ataset|[D|d]omain", variable =
    "[V|v]ariable [[N|n]ame]?|[V|v]ariables?", key_seq =
    "[V|v]ariable [O|o]rder|[S|s]eq|[O|o]rder", keep = "[K|k]eep|[M|m]andatory"),
  sheet = "[V|v]ar"
)
}
\arguments{
\item{doc}{Named list of datasets @seealso \code{\link[=read_all_sheets]{read_all_sheets()}} for exact
format}

\item{cols}{Named vector of column names. The column names can be regular
expressions for more flexibility. But, the names must follow the given pattern}

\item{sheet}{Regular expression for the sheet name}
}
\value{

}
\description{
Creates the ds_vars from a list of datasets (optionally filtered by the sheet
input). The named vector \code{cols} is used to determine which is the correct
sheet and renames the columns
}
\seealso{
Other spec builder: 
\code{\link{spec_type_to_codelist}()},
\code{\link{spec_type_to_derivations}()},
\code{\link{spec_type_to_ds_spec}()},
\code{\link{spec_type_to_value_spec}()},
\code{\link{spec_type_to_var_spec}()}
}
\concept{spec builder}
