% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_builder.R
\name{spec_type_to_var_spec}
\alias{spec_type_to_var_spec}
\title{Spec to var_spec}
\usage{
spec_type_to_var_spec(doc,
  cols = c(
    "variable" = "[N|n]ame|[V|v]ariables?",
    "length"   = "[L|l]ength",
    "label"    = "[L|l]abel",
    "type"     = "[T|t]ype",
    "dataset"  = "[D|d]ataset|[D|d]omain",
    "format"   = "[F|f]ormat"
  ),
  sheet = "[V|v]ar")
}
\arguments{
\item{doc}{Named list of datasets @seealso \code{\link[=read_all_sheets]{read_all_sheets()}} for exact
format}

\item{cols}{Named vector of column names. The column names can be regular
expressions for more flexibility. But, the names must follow the given pattern}

\item{sheet}{Regular expression for the sheet name}
}
\value{
a dataset formatted for the metacore object
}
\description{
Creates the var_spec from a list of datasets (optionally filtered by the sheet
input). The named vector \code{cols} is used to determine which is the correct
sheet and renames the columns. (Note: the keep column will be converted logical)
}
\seealso{
Other spec builders: 
\code{\link{spec_type_to_codelist}()},
\code{\link{spec_type_to_comments}()},
\code{\link{spec_type_to_derivations}()},
\code{\link{spec_type_to_documents}()},
\code{\link{spec_type_to_ds_spec}()},
\code{\link{spec_type_to_ds_vars}()},
\code{\link{spec_type_to_supp}()},
\code{\link{spec_type_to_value_spec}()}
}
\concept{spec builders}
