% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_builders.R
\name{xml_to_derivations}
\alias{xml_to_derivations}
\title{XML to derivation table}
\usage{
xml_to_derivations(doc)
}
\arguments{
\item{doc}{?}
}
\value{
dataframe with derivation id's and derivations
}
\description{
This reads in a xml document and gets all the derivations/comments. These can
be cross referenced to variables using the derivation_id's
}
\seealso{
Other xml builder: 
\code{\link{xml_to_codelist}()},
\code{\link{xml_to_ds_spec}()},
\code{\link{xml_to_ds_vars}()},
\code{\link{xml_to_value_spec}()},
\code{\link{xml_to_var_spec}()}
}
\concept{xml builder}
