% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_builders.R
\name{xml_to_ds_spec}
\alias{xml_to_ds_spec}
\title{XML to Data Set Spec}
\usage{
xml_to_ds_spec(doc)
}
\arguments{
\item{doc}{xml document}
}
\value{
data frame with the data set specifications
}
\description{
Creates a dataset specification, which has the domain name and label for each dataset
}
\seealso{
Other xml builder: 
\code{\link{xml_to_codelist}()},
\code{\link{xml_to_derivations}()},
\code{\link{xml_to_ds_vars}()},
\code{\link{xml_to_value_spec}()},
\code{\link{xml_to_var_spec}()}
}
\concept{xml builder}
