% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_builders.R
\name{xml_to_value_spec}
\alias{xml_to_value_spec}
\title{XML to value spec}
\usage{
xml_to_value_spec(doc)
}
\arguments{
\item{doc}{xml document}
}
\value{
tibble with the value level information
}
\description{
Takes a define xml and pulls out the value level metadata including codelist_id's,
defines_id's, and where clause. There is one row per variable expect when there
is a where clause, at which point there is one row per value.
}
\seealso{
Other xml builder: 
\code{\link{xml_to_codelist}()},
\code{\link{xml_to_derivations}()},
\code{\link{xml_to_ds_spec}()},
\code{\link{xml_to_ds_vars}()},
\code{\link{xml_to_var_spec}()}
}
\concept{xml builder}
